Berechtigungen hinzugefügt

+++ vilesci/stammdaten/reihungstest_administration.php
@@ -123,11 +123,14 @@ echo '<!DOCTYPE HTML PUBLIC "-//W3C//DTD HTML 4.01 Transitional//DE" "http://www
 	<body class="Background_main">
 	<h2>Reihungstest - Administration</h2>';
 
-if(!$rechte->isBerechtigt('basis/testtool', null, 'suid'))
-		die('Sie haben keine Berechtigung fuer diese Seite');
+if(!$rechte->isBerechtigt('basis/testtool', null, 's'))
+		die('Sie haben keine Berechtigung für diese Seite');
 
 if(isset($_POST['personzuteilen']))
 {
+	if(!$rechte->isBerechtigt('basis/testtool', null, 'sui'))
+		die('<span class="error">Sie haben keine Berechtigung für diese Aktion. <a href="reihungstest_administration.php">Seite neu laden</a></span>');
+	
 	$prestudent = new prestudent();
 	if($prestudent->load($_POST['prestudent_id']))
 	{
@@ -252,6 +255,9 @@ if($result = $db->db_query($qry))
 echo '<input type="submit" value="Dieses Teilgebiet l&ouml;schen" name="deleteteilgebiet" onclick="return confirm(\'Antworten dieses Gebietes wirklich löschen?\')">&nbsp;&nbsp;&nbsp;&nbsp;';
 if(isset($_POST['deleteteilgebiet']))
 {
+	if(!$rechte->isBerechtigt('basis/testtool', null, 'suid'))
+		die('<span class="error">Sie haben keine Berechtigung für diese Aktion. <a href="reihungstest_administration.php">Seite neu laden</a></span>');
+	
 	if(isset($_POST['prestudent']) && isset($_POST['gebiet']) && 
 	   is_numeric($_POST['prestudent']) && is_numeric($_POST['gebiet']))
 	{
@@ -365,6 +371,9 @@ echo '<hr><br><a href="'.$_SERVER['PHP_SELF'].'?action=deletedummyanswers" oncli
 
 if(isset($_GET['action']) && $_GET['action']=='deletedummyanswers')
 {
+	if(!$rechte->isBerechtigt('basis/testtool', null, 'suid'))
+		die('<span class="error">Sie haben keine Berechtigung für diese Aktion. <a href="reihungstest_administration.php">Seite neu laden</a></span>');
+		
 	$qry = "DELETE FROM testtool.tbl_antwort WHERE pruefling_id=841;
 			DELETE FROM testtool.tbl_pruefling_frage where pruefling_id=841;";
 	if($db->db_query($qry))
@@ -377,6 +386,9 @@ if(isset($_GET['action']) && $_GET['action']=='deletedummyanswers')
 echo '<br><br>';
 if(isset($_POST['savedummystg']) && isset($_POST['stg']))
 {
+	if(!$rechte->isBerechtigt('basis/testtool', null, 'su'))
+		die('<span class="error">Sie haben keine Berechtigung für diese Aktion. <a href="reihungstest_administration.php">Seite neu laden</a></span>');
+		
 	$qry = "UPDATE public.tbl_prestudent SET studiengang_kz=".$db->db_add_param($_POST['stg'])." WHERE prestudent_id='13478';
 	UPDATE testtool.tbl_pruefling SET studiengang_kz=".$db->db_add_param($_POST['stg'])." WHERE prestudent_id='13478';";	
 	if($db->db_query($qry))
@@ -627,6 +639,9 @@ echo 'Freigeschaltene Reihungstests:';
 
 if(isset($_GET['action']) && $_GET['action']=='sperren')
 {
+	if(!$rechte->isBerechtigt('basis/testtool', null, 'su'))
+		die('<span class="error">Sie haben keine Berechtigung für diese Aktion. <a href="reihungstest_administration.php">Seite neu laden</a></span>');
+		
 	$rt = new reihungstest();
 	if($rt->load($_GET['reihungstest_id']))
 	{
